big improvements
blurry png problem solved
bad text rendering problem solved
popup in social media - custom in each design
some elements in photo booth become custom
changed over yellow element in settings to .png (was gl element)
some bug fixed

+++ Engine/src/tools/Utils.h
@@ -49,7 +49,7 @@ class Utils
 
 		static void  drawGraphicsToFBO(Fbo fbo, const function<void ()>& graphicsFunc);
 		static ci::gl::Texture drawGraphicsToFBO(int width, int height, const function<void ()>& graphicsFunc);
-		static ci::gl::Texture drawGraphicsToFBO(ci::Vec2f size, const function<void ()>& graphicsFunc);
+		static ci::gl::Texture drawGraphicsToFBO(ci::Vec2i size, const function<void ()>& graphicsFunc);
 
 		static void clearFBO(Fbo fbo);
 		static void printVideoMemoryInfo();
